+++ packages/building-webpack/modern-and-legacy-config.js
@@ -62,7 +62,11 @@ function createConfig(options, legacy) {
                       'lit-html': ['html'],
                       'lit-element': ['html', { name: 'css', encapsulation: 'style' }],
                     },
-                    htmlMinifier: { collapseWhitespace: true, removeComments: true },
+                    htmlMinifier: {
+                      collapseWhitespace: true,
+                      removeComments: true,
+                      caseSensitive: true,
+                    },
                   },
                 ],
                 // webpack does not support import.meta.url yet, so we rewrite them in babel