+++ packages/create/src/generators/building-rollup/index.js
@@ -1,12 +1,7 @@
-import { askTagInfo } from '../../helpers.js';
-
 const BuildingRollupMixin = subclass =>
   class extends subclass {
     async execute() {
       await super.execute();
-
-      const { tagName, className } = await askTagInfo();
-      this.templateData = { ...this.templateData, tagName, className };
 
       this.copyTemplateJsonInto(
         `${__dirname}/templates/_package.json`,