+++ packages/building-rollup/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.3.0...@open-wc/building-rollup@0.3.1) (2019-04-14)
+
+
+### Bug Fixes
+
+* update generator usage ([5d284d4](https://github.com/open-wc/open-wc/commit/5d284d4))
+
+
+
+
+
 # [0.3.0](https://github.com/open-wc/open-wc/compare/@open-wc/building-rollup@0.2.4...@open-wc/building-rollup@0.3.0) (2019-04-08)