+++ packages/testing/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.4](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.11.3...@open-wc/testing@0.11.4) (2019-04-14)
+
+
+### Bug Fixes
+
+* update generator usage ([5d284d4](https://github.com/open-wc/open-wc/commit/5d284d4))
+
+
+
+
+
 ## [0.11.3](https://github.com/open-wc/open-wc/compare/@open-wc/testing@0.11.2...@open-wc/testing@0.11.3) (2019-04-13)
 
 **Note:** Version bump only for package @open-wc/testing